modulo-05-compose/modulo-05-compose/projeto-taskmanager/requirements.txt
Flask
python-dotenv
psycopg2-binary
redis
gunicorn
